requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==19.2.0
certifi==2017.4.17
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.0.1
itsdangerous==0.24
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.8.2
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.21.1
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.4.0
